Move scene class
