+++ tests/Scene_Test.cpp
@@ -4,7 +4,7 @@
 #include <gtest/gtest.h>
 
 TEST(Scene, HierarchyIteration) {
-	arch::Scene scene;
+	arch::scene::Scene scene;
 
 	auto e = std::array<arch::hier::HierarchyNode*, 11>();
 
@@ -25,16 +25,20 @@ TEST(Scene, HierarchyIteration) {
 	{ e[8]->setParent(*e[7]); }
 	e[9]->setParent(*e[0]);
 
-	ASSERT_TRUE(std::ranges::equal(
-		e,
-		scene.rootNode().recursiveIterable(),
-		{},
-		[](const arch::hier::HierarchyNode* node) { return node->entity(); }
-	));
-	ASSERT_TRUE(std::ranges::equal(
-		std::views::reverse(e),
-		std::views::reverse(scene.rootNode().recursiveIterable()),
-		{},
-		[](const arch::hier::HierarchyNode* node) { return node->entity(); }
-	));
+	ASSERT_TRUE(
+		std::ranges::equal(
+			e,
+			scene.rootNode().recursiveIterable(),
+			{},
+			[](const arch::hier::HierarchyNode* node) { return node->entity(); }
+		)
+	);
+	ASSERT_TRUE(
+		std::ranges::equal(
+			std::views::reverse(e),
+			std::views::reverse(scene.rootNode().recursiveIterable()),
+			{},
+			[](const arch::hier::HierarchyNode* node) { return node->entity(); }
+		)
+	);
 }
